% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durbanplasticwaste-package.R
\docType{package}
\name{durbanplasticwaste-package}
\alias{durbanplasticwaste}
\alias{durbanplasticwaste-package}
\title{durbanplasticwaste: Durban (South Aftica) Plastic Waste Data}
\description{
Waste characterisation data for extended producer responsibility. The plastic waste was collected in the uMngeni River system in Durban, South Africa, and categorised by brand, corporate group, application, and type of plastic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://global-health-engineering.github.io/durbanplasticwaste/}
}

}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Raúl Bergen \email{raul.bergen@me.com} (\href{https://orcid.org/0000-0002-0678-6425}{ORCID})
  \item Chiara Meyer-Piening \email{first.last@example.com} (\href{https://orcid.org/0000-0003-2368-7418}{ORCID})
}

Other contributors:
\itemize{
  \item Boynton Lin \email{lbonyton@ethz.ch} (\href{https://orcid.org/0000-0002-9944-3834}{ORCID}) [contributor]
  \item Elizabeth Tilley \email{tilleye@ethz.ch} (\href{https://orcid.org/0000-0002-2095-9724}{ORCID}) [contributor]
  \item Marc Kalina \email{mkalina@ethz.ch} (\href{https://orcid.org/0000-0002-6335-3845}{ORCID}) [contributor]
  \item Siphiwe Rakgabale \email{siphiwerakgabale@gmail.com} (\href{https://orcid.org/0009-0007-3253-6327}{ORCID}) [contributor]
  \item Sfiso Nduduzo Luvuno \email{sfisonduduzo.luvuno@gmail.com} [contributor]
  \item Global Health Engineering [funder]
}

}
\keyword{internal}
