% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterboom_counts.R
\docType{data}
\name{litterboom_counts}
\alias{litterboom_counts}
\title{Plastic waste litterboom counts}
\format{
A tibble with 2784 rows and 7 variables:
\describe{
\item{date}{Date of the collected litterboom sample.}
\item{location}{Descriptive name of the sample location. See \code{\link{locations}}
for longitude and latitude .}
\item{brand}{Brand name of the collected item (e.g. Coca Cola).}
\item{group}{Group name that owns the brand (e.g. Coca Cola Beverages
South Africa}
\item{plastic}{Type of plastic of the item. Identified plastic types are
PET, HDPE, and PP. HDPE and PP were categorised together as HDPE/PP.}
\item{category}{Categorisation of waste into 15 product type categories
(e.g. Alcohol, Milk, Tobacco, Water).}
\item{count}{Number of counted items.}
}
}
\usage{
litterboom_counts
}
\description{
Count of individual plastic waste items categorised by brand, corporate
group, plastic type, and usage application. Waste was collected on 14 days
in 6 locations in Durban, South Africa.
}
\keyword{datasets}
