% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterboom_weights.R
\docType{data}
\name{litterboom_weights}
\alias{litterboom_weights}
\title{Plastic waste litterboom weights}
\format{
A tibble with 14 rows and 3 variables:
\describe{
\item{date}{Date of the collected litterboom sample.}
\item{location}{Descriptive name of the sample location. See \code{\link{locations}}
for longitude and latitude .}
\item{pet}{Weight (in kg) of PET items.}
\item{hdpe_pp}{Weight (in kg) of HDPE and PP items.}
}
}
\usage{
litterboom_weights
}
\description{
Weight in kg of identified PET and HDPE/PP items.
}
\keyword{datasets}
