% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{Plastic waste litterboom sampling locations}
\format{
A tibble with 6 rows and 3 variables:
\describe{
\item{location}{Descriptive name of the sample location.}
\item{latitude}{Latitude coordinate.}
\item{longitude}{Longitude coordinate.}
}
}
\usage{
locations
}
\description{
Plastic waste litterboom sampling locations
}
\keyword{datasets}
